cp can be float

+++ cogs/admin.py
@@ -546,7 +546,7 @@ class Admin(commands.Cog, name="Admin"):
     
     @commands.command()
     @commands.has_any_role("Mod Friend")
-    async def xfer(self, ctx, charName, level : int, cp :int, user, items = ""):
+    async def xfer(self, ctx, charName, level : int, cp :float, user, items = ""):
         msg = ctx.message
         rewardList = msg.raw_mentions
         rewardUser = ""
